multiple files template
