+++ templates/cw721-template/test_generate.sh
